refactor: unify throughput query logic in shared package
- Create internal/pkg/db/throughput.go with unified BuildThroughputQuery() and
  BuildProbeStatsQuery() functions
- Extract helper functions from auto-generated dashboard.resolvers.go
- Replace MAX(id) subquery in channel_probe.go with shared query builder
- Support both ROW_NUMBER() (PostgreSQL) and MAX(id) (SQLite fallback) modes
- Maintain same query semantics and confidence calculation logic

Addresses code review comments:
- Fixes query logic inconsistency between channel_probe and dashboard resolvers
- Moves helpers out of gqlgen auto-generated file
- No changes to test behavior (all tests pass)
