+++ internal/server/biz/channel_probe.go
@@ -17,6 +17,7 @@ import (
 	"github.com/looplj/axonhub/internal/ent/channelprobe"
 	"github.com/looplj/axonhub/internal/ent/privacy"
 	"github.com/looplj/axonhub/internal/log"
+	"github.com/looplj/axonhub/internal/pkg/db"
 	"github.com/looplj/axonhub/internal/pkg/xtime"
 )
 
@@ -140,14 +141,14 @@ func (svc *ChannelProbeService) computeAllChannelProbeStats(
 	}
 
 	dbDriver := svc.db.Driver()
-	db, ok := dbDriver.(*entsql.Driver)
+	sqlDB, ok := dbDriver.(*entsql.Driver)
 	if !ok {
 		return nil, fmt.Errorf("failed to get underlying SQL driver")
 	}
 
 	// Detect dialect to use appropriate placeholder syntax
 	// PostgreSQL uses $1, $2, etc. while SQLite uses ? placeholders
-	dialectName := db.Dialect()
+	dialectName := sqlDB.Dialect()
 	useDollarPlaceholders := dialectName == dialect.Postgres
 
 	// Build args slice for parameterized query
@@ -169,68 +170,13 @@ func (svc *ChannelProbeService) computeAllChannelProbeStats(
 		channelIDFilter = fmt.Sprintf("AND se.channel_id IN (%s)", strings.Join(placeholders, ","))
 	}
 
-	// Build query with dialect-aware timestamp placeholders
-	// Use a subquery approach that avoids window functions for better SQLite compatibility
-	var query string
-	if useDollarPlaceholders {
-		query = fmt.Sprintf(`
-SELECT 
-    se.channel_id,
-    COUNT(*) as total_count,
-    COUNT(*) as success_count,
-    SUM(ul.completion_tokens + COALESCE(ul.completion_reasoning_tokens, 0) + COALESCE(ul.completion_audio_tokens, 0)) as total_tokens,
-    SUM(CASE WHEN se.stream AND se.metrics_first_token_latency_ms IS NOT NULL
-         THEN CASE WHEN se.metrics_first_token_latency_ms >= se.metrics_latency_ms
-              THEN 0
-              ELSE se.metrics_latency_ms - se.metrics_first_token_latency_ms END
-         ELSE se.metrics_latency_ms END) as effective_latency_ms,
-    SUM(se.metrics_first_token_latency_ms) as total_first_token_latency,
-    COUNT(DISTINCT se.request_id) as request_count
-FROM request_executions se
-JOIN usage_logs ul ON se.request_id = ul.request_id
-WHERE se.status = 'completed'
-    AND se.metrics_latency_ms > 0
-    AND se.created_at >= $1
-    AND se.created_at < $2
-    AND se.id = (
-        SELECT MAX(re2.id)
-        FROM request_executions re2
-        WHERE re2.request_id = se.request_id
-    )
-    %s
-GROUP BY se.channel_id
-ORDER BY se.channel_id`, channelIDFilter)
-	} else {
-		query = fmt.Sprintf(`
-SELECT 
-    se.channel_id,
-    COUNT(*) as total_count,
-    COUNT(*) as success_count,
-    SUM(ul.completion_tokens + COALESCE(ul.completion_reasoning_tokens, 0) + COALESCE(ul.completion_audio_tokens, 0)) as total_tokens,
-    SUM(CASE WHEN se.stream AND se.metrics_first_token_latency_ms IS NOT NULL
-         THEN CASE WHEN se.metrics_first_token_latency_ms >= se.metrics_latency_ms
-              THEN 0
-              ELSE se.metrics_latency_ms - se.metrics_first_token_latency_ms END
-         ELSE se.metrics_latency_ms END) as effective_latency_ms,
-    SUM(se.metrics_first_token_latency_ms) as total_first_token_latency,
-    COUNT(DISTINCT se.request_id) as request_count
-FROM request_executions se
-JOIN usage_logs ul ON se.request_id = ul.request_id
-WHERE se.status = 'completed'
-    AND se.metrics_latency_ms > 0
-    AND se.created_at >= ?
-    AND se.created_at < ?
-    AND se.id = (
-        SELECT MAX(re2.id)
-        FROM request_executions re2
-        WHERE re2.request_id = se.request_id
-    )
-    %s
-GROUP BY se.channel_id
-ORDER BY se.channel_id`, channelIDFilter)
+	queryMode := db.ThroughputModeROW_NUMBER
+	if !useDollarPlaceholders {
+		queryMode = db.ThroughputModeMaxID
 	}
+	query := db.BuildProbeStatsQuery(useDollarPlaceholders, channelIDFilter, queryMode)
 
-	rows, err := db.DB().QueryContext(ctx, query, args...)
+	rows, err := sqlDB.DB().QueryContext(ctx, query, args...)
 	if err != nil {
 		return nil, fmt.Errorf("failed to query channel probe stats: %w", err)
 	}
