+++ internal/server/gql/dashboard.resolvers.go
@@ -25,6 +25,7 @@ import (
 	"github.com/looplj/axonhub/internal/ent/usagelog"
 	"github.com/looplj/axonhub/internal/log"
 	"github.com/looplj/axonhub/internal/objects"
+	"github.com/looplj/axonhub/internal/pkg/db"
 	"github.com/looplj/axonhub/internal/pkg/xtime"
 	"github.com/looplj/axonhub/internal/scopes"
 )
@@ -791,14 +792,14 @@ func (r *queryResolver) FastestChannels(ctx context.Context, input FastestChanne
 	}
 	var results []channelStats
 	dbDriver := r.client.Driver()
-	db, ok := dbDriver.(*sql.Driver)
+	sqlDB, ok := dbDriver.(*sql.Driver)
 	if !ok {
 		return nil, fmt.Errorf("failed to get underlying SQL driver")
 	}
 
 	// Detect dialect to use appropriate placeholder syntax
 	// PostgreSQL uses $1, $2, etc. while SQLite uses ? placeholders
-	dialectName := db.Dialect()
+	dialectName := sqlDB.Dialect()
 	useDollarPlaceholders := dialectName == dialect.Postgres
 
 	// Build query using shared helper function
@@ -807,15 +808,16 @@ func (r *queryResolver) FastestChannels(ctx context.Context, input FastestChanne
 	if sqlLimit < 20 {
 		sqlLimit = 20
 	}
-	query := buildThroughputQuery(
+	query := db.BuildThroughputQuery(
 		useDollarPlaceholders,
-		ThroughputQueryByChannel,
+		db.ThroughputQueryByChannel,
 		sqlLimit,
+		db.ThroughputModeROW_NUMBER,
 	)
 
 	// Use UTC for the time parameter to match the timezone of the created_at column.
 	// This assumes created_at is stored in UTC, which is consistent with the application's timezone handling.
-	rows, err := db.DB().QueryContext(ctx, query, since.UTC())
+	rows, err := sqlDB.DB().QueryContext(ctx, query, since.UTC())
 	if err != nil {
 		return nil, fmt.Errorf("failed to query fastest channels: %w", err)
 	}
@@ -871,7 +873,7 @@ func (r *queryResolver) FastestChannels(ctx context.Context, input FastestChanne
 	}
 
 	scoredResults := lo.Map(results, func(item channelStats, _ int) scoredChannelItem {
-		conf := calculateConfidenceLevel(int(item.RequestCount), median)
+		conf := db.CalculateConfidenceLevel(int(item.RequestCount), median)
 		score := 0
 		switch conf {
 		case "high":
@@ -972,14 +974,14 @@ func (r *queryResolver) FastestModels(ctx context.Context, input FastestChannels
 	}
 	var results []modelStats
 	dbDriver := r.client.Driver()
-	db, ok := dbDriver.(*sql.Driver)
+	sqlDB, ok := dbDriver.(*sql.Driver)
 	if !ok {
 		return nil, fmt.Errorf("failed to get underlying SQL driver")
 	}
 
 	// Detect dialect to use appropriate placeholder syntax
 	// PostgreSQL uses $1, $2, etc. while SQLite uses ? placeholders
-	dialectName := db.Dialect()
+	dialectName := sqlDB.Dialect()
 	useDollarPlaceholders := dialectName == dialect.Postgres
 
 	// Build query with dialect-aware timestamp placeholder
@@ -988,13 +990,14 @@ func (r *queryResolver) FastestModels(ctx context.Context, input FastestChannels
 	if sqlLimit < 20 {
 		sqlLimit = 20
 	}
-	query := buildThroughputQuery(
+	query := db.BuildThroughputQuery(
 		useDollarPlaceholders,
-		ThroughputQueryByModel,
+		db.ThroughputQueryByModel,
 		sqlLimit,
+		db.ThroughputModeROW_NUMBER,
 	)
 
-	rows, err := db.DB().QueryContext(ctx, query, since.UTC())
+	rows, err := sqlDB.DB().QueryContext(ctx, query, since.UTC())
 	if err != nil {
 		return nil, fmt.Errorf("failed to query fastest models: %w", err)
 	}
@@ -1049,7 +1052,7 @@ func (r *queryResolver) FastestModels(ctx context.Context, input FastestChannels
 	}
 
 	scoredResults := lo.Map(results, func(item modelStats, _ int) scoredModelItem {
-		conf := calculateConfidenceLevel(int(item.RequestCount), median)
+		conf := db.CalculateConfidenceLevel(int(item.RequestCount), median)
 		score := 0
 		switch conf {
 		case "high":
@@ -1112,129 +1115,3 @@ func (r *queryResolver) FastestModels(ctx context.Context, input FastestChannels
 //   - When renaming or deleting a resolver the old code will be put in here. You can safely delete
 //     it when you're done.
 //   - You have helper methods in this file. Move them out to keep these resolver files clean.
-
-// ThroughputQueryType identifies the type of throughput query to build.
-// This enum ensures only predefined, validated query patterns can be used.
-type ThroughputQueryType int
-
-const (
-	// ThroughputQueryByChannel groups throughput statistics by channel.
-	// Uses channels table for channel metadata.
-	ThroughputQueryByChannel ThroughputQueryType = iota
-	// ThroughputQueryByModel groups throughput statistics by model.
-	// Uses requests and models tables for model metadata.
-	ThroughputQueryByModel
-)
-
-// queryFragmentConfig holds the SQL fragments for a specific query type.
-// These fragments are predefined constants and never accept user input.
-type queryFragmentConfig struct {
-	selectColumns string
-	joinClause    string
-	groupBy       string
-}
-
-// allowedQueryConfigs maps each ThroughputQueryType to its validated SQL fragments.
-// This allowlist ensures only safe, pre-approved SQL patterns can be executed.
-var allowedQueryConfigs = map[ThroughputQueryType]queryFragmentConfig{
-	ThroughputQueryByChannel: {
-		selectColumns: "se.channel_id,\n    c.name as channel_name,\n    c.type as channel_type,",
-		joinClause:    "JOIN channels c ON se.channel_id = c.id",
-		groupBy:       "se.channel_id, c.name, c.type",
-	},
-	ThroughputQueryByModel: {
-		selectColumns: "r.model_id,\n    m.name as model_name,",
-		joinClause:    "JOIN requests r ON se.request_id = r.id\nJOIN models m ON r.model_id = m.model_id",
-		groupBy:       "r.model_id, m.name",
-	},
-}
-
-// buildThroughputQuery constructs a SQL query for throughput statistics.
-// SECURITY NOTE: This function now uses ThroughputQueryType enum instead of raw SQL strings.
-// The SQL fragments are retrieved from a predefined allowlist (allowedQueryConfigs),
-// eliminating the risk of SQL injection from user input. Only the queryType parameter
-// determines which SQL pattern is used, and limit is validated as a positive integer.
-//
-// COMPATIBILITY NOTE: This query uses ROW_NUMBER() window function which requires
-// SQLite 3.25+ (released 2018-09-15). All supported database dialects (PostgreSQL,
-// MySQL 8.0+, TiDB, SQLite 3.25+) support this function.
-func buildThroughputQuery(useDollarPlaceholders bool, queryType ThroughputQueryType, limit int) string {
-	// Validate that limit is positive to prevent malformed queries
-	if limit <= 0 {
-		limit = 20 // Default fallback
-	}
-
-	// Retrieve the validated query configuration from the allowlist
-	config, ok := allowedQueryConfigs[queryType]
-	if !ok {
-		// This should never happen with proper enum usage, but return a safe default
-		// that will result in an empty result set rather than a malformed query
-		config = allowedQueryConfigs[ThroughputQueryByChannel]
-	}
-
-	placeholder := "$1"
-	if !useDollarPlaceholders {
-		placeholder = "?"
-	}
-
-	return `
-WITH successful_execs AS (
-    SELECT
-        request_id,
-        channel_id,
-        metrics_latency_ms,
-        metrics_first_token_latency_ms,
-        stream,
-        ROW_NUMBER() OVER (PARTITION BY request_id ORDER BY created_at DESC) as rn
-    FROM request_executions
-    WHERE status = 'completed' AND metrics_latency_ms > 0 AND created_at >= ` + placeholder + `
-)
-SELECT
-    ` + config.selectColumns + `
-    SUM(ul.completion_tokens + COALESCE(ul.completion_reasoning_tokens, 0) + COALESCE(ul.completion_audio_tokens, 0)) as tokens_count,
-    SUM(se.metrics_latency_ms) as latency_ms,
-    COUNT(DISTINCT se.request_id) as request_count,
-    CASE
-        WHEN SUM(CASE WHEN se.stream AND se.metrics_first_token_latency_ms IS NOT NULL
-                 THEN se.metrics_latency_ms - se.metrics_first_token_latency_ms
-                 ELSE se.metrics_latency_ms END) > 0
-        THEN SUM(ul.completion_tokens + COALESCE(ul.completion_reasoning_tokens, 0) + COALESCE(ul.completion_audio_tokens, 0)) * 1000.0
-             / SUM(CASE WHEN se.stream AND se.metrics_first_token_latency_ms IS NOT NULL
-                   THEN se.metrics_latency_ms - se.metrics_first_token_latency_ms
-                   ELSE se.metrics_latency_ms END)
-        ELSE 0
-    END as throughput
-FROM successful_execs se
-JOIN usage_logs ul ON se.request_id = ul.request_id
-` + config.joinClause + `
-WHERE se.rn = 1
-GROUP BY ` + config.groupBy + `
-ORDER BY throughput DESC
-LIMIT ` + fmt.Sprintf("%d", limit)
-}
-
-func calculateConfidenceLevel(requestCount int, median float64) string {
-	// When median is 0, we cannot calculate a meaningful ratio (requestCount/median),
-	// so we default to low confidence since we lack sufficient data for reliable inference.
-	if median == 0 {
-		return "low"
-	}
-
-	// Absolute minimum request thresholds for confidence levels
-	// These ensure items with very few requests are always low confidence
-	const minRequestsForMedium = 100
-	const minRequestsForHigh = 500
-
-	if requestCount < minRequestsForMedium {
-		return "low"
-	}
-
-	ratio := float64(requestCount) / median
-	if ratio >= 1.5 && requestCount >= minRequestsForHigh {
-		return "high"
-	}
-	if ratio >= 0.5 {
-		return "medium"
-	}
-	return "low"
-}
